test: check all route test
